backend/requirements.txt
fastapi
uvicorn[standard]
redis
python-multipart
PyPDF2
python-docx
sentence-transformers
numpy
scikit-learn
google-generativeai
python-dotenv
pydantic
pydantic-settings
httpx
beautifulsoup4
lxml
asyncio
python-json-logger
scalar-fastapi
aiohttp